WIP (still have to update tests)
addMemberAs and DeleteMemberAs will now be able to add/delete members of a Grouping "Preemptively"

addMemberToGrouping and deleteMemberFromGrouping will remove any preemptions while adding or deleting members.

+++ src/main/java/edu/hawaii/its/api/service/GroupingsServiceImpl.java
@@ -149,7 +149,13 @@ public class GroupingsServiceImpl implements GroupingsService {
 
 
     @Override
-    public List<GroupingsServiceResult> addGrouping(String username, String path, List<String> basis, List<String> include, List<String> exclude, List<String> owners) {
+    public List<GroupingsServiceResult> addGrouping(
+            String username,
+            String path,
+            List<String> basis,
+            List<String> include,
+            List<String> exclude,
+            List<String> owners) {
 
         //This method will not work until Grouper is updated
 
@@ -193,11 +199,11 @@ public class GroupingsServiceImpl implements GroupingsService {
 //            addGroupingResults.add(updateLastModified(path));
 //
 //            for (Map.Entry<String, List<String>> entry : memberLists.entrySet()) {
-//                addGroupingResults.add(addMemberAs(username, path + entry.getKey(), entry.getValue()));
+//                addGroupingResults.add(addMembersAs(username, path + entry.getKey(), entry.getValue()));
 //                addGroupingResults.add(updateLastModified(path + entry.getKey()));
 //            }
 //
-//            addGroupingResults.add(addMemberAs(username, GROUPING_OWNERS, memberLists.get(OWNERS)));
+//            addGroupingResults.add(addMembersAs(username, GROUPING_OWNERS, memberLists.get(OWNERS)));
 //            addGroupingResults.add(updateLastModified(GROUPING_OWNERS));
 //
 //        } else {
@@ -464,7 +470,7 @@ public class GroupingsServiceImpl implements GroupingsService {
 
         if (groupOptInPermission(username, addGroup)) {
             results.add(deleteMemberAs(username, deleteGroup, username));
-            results.add(addMemberAs(username, addGroup, username));
+            results.addAll(addMemberAs(username, addGroup, username));
 
             if (inGroup(addGroup, username)) {
                 results.add(addSelfOpted(addGroup, username));
@@ -1145,77 +1151,113 @@ public class GroupingsServiceImpl implements GroupingsService {
         return makeGroupingsServiceResult("FAILURE: " + username + " is not an admin", action);
     }
 
-    /**
-     * @param username:  username of owner adding member
-     * @param group:     path to group the user to be added will be added to
-     * @param userToAdd: username of user to be added to group
-     * @return information about success of action
-     */
     @Override
-    public GroupingsServiceResult addMemberAs(String username, String group, String userToAdd) {
-        logger.info("addMemberAs; user: " + username + "; group: " + group + "; userToAdd: " + userToAdd + ";");
+    public List<GroupingsServiceResult> addMemberAs(String username, String group, String userToAdd) {
+        logger.info("addMemberToGroup; user: " + username + "; group: " + group + "; userToAdd: " + userToAdd + ";");
 
-        WsSubjectLookup user = gf.makeWsSubjectLookup(username);
+        //todo change to map? (add, delete) or (include, exclude, owners)
+        List<GroupingsServiceResult> gsrs = new ArrayList<>();
         String action = "add users to " + group;
-        String composite = parentGroupingPath(group);
-        String basis = composite + BASIS;
-        String exclude = composite + EXCLUDE;
-        String include = composite + INCLUDE;
 
-        if (group.endsWith(INCLUDE)) {
-            //if they are in the exclude, get them out
-            if (inGroup(exclude, userToAdd)) {
-                gf.makeWsDeleteMemberResults(
-                        exclude,
-                        user,
-                        userToAdd);
+        if (isOwner(parentGroupingPath(group), username) || isSuperuser(username)) {
+            WsSubjectLookup user = gf.makeWsSubjectLookup(username);
+            String composite = parentGroupingPath(group);
+            String exclude = composite + EXCLUDE;
+            String include = composite + INCLUDE;
+            String owners = composite + OWNERS;
 
-                updateLastModified(exclude);
-                updateLastModified(composite);
-            }
-            //only add them to the include if they are not in the basis
-            if (!inGroup(basis, userToAdd)) {
-                WsAddMemberResults addMemberResults = gf.makeWsAddMemberResults(include, user, userToAdd);
+            boolean updateComposite = false;
+            boolean updateExclude = false;
+            boolean updateInclude = false;
+            boolean updateOwners = false;
 
-                updateLastModified(include);
-                updateLastModified(composite);
+            //check to see if it is the include, exclude or owners
+            if (group.endsWith(INCLUDE)) {
+                //check to see if userToAdd is already in include
+                if (!inGroup(include, userToAdd)) {
+                    //if userToAdd is in exclude, get them out
+                    if (inGroup(exclude, username)) {
+                        gf.makeWsDeleteMemberResults(
+                                exclude,
+                                user,
+                                userToAdd);
 
-                return makeGroupingsServiceResult(addMemberResults, action);
-            } else {
-                return makeGroupingsServiceResult(SUCCESS + ": " + userToAdd + " was in basis", action);
-            }
-        } else if (group.endsWith(EXCLUDE)) {
-            //if they are in the include, get them out
-            if (inGroup(include, userToAdd)) {
-                gf.makeWsDeleteMemberResults(
-                        include,
-                        user,
-                        userToAdd);
+                        updateExclude = true;
+                    }
+                    //add to include
+                    WsAddMemberResults addMemberResults = gf.makeWsAddMemberResults(include, user, userToAdd);
 
-                updateLastModified(include);
-                updateLastModified(composite);
+                    updateInclude = true;
+
+                    gsrs.add(makeGroupingsServiceResult(addMemberResults, action));
+                }
+                //They are already in the group, so just return SUCCESS
+                gsrs.add(makeGroupingsServiceResult(SUCCESS + ": " + userToAdd + " was already in " + group, action));
             }
 
-            //only add them to the exclude if they are in the basis
-            if (inGroup(basis, userToAdd)) {
-                WsAddMemberResults addMemberResults = gf.makeWsAddMemberResults(exclude, user, userToAdd);
+            //if exclude check if userToAdd is in the include
+            else if (group.endsWith(EXCLUDE)) {
+                //check to see if userToAdd is already in exclude
+                if (!inGroup(exclude, userToAdd)) {
+                    //if userToAdd is in include, get them out
+                    if (inGroup(include, username)) {
+                        gf.makeWsDeleteMemberResults(
+                                include,
+                                user,
+                                userToAdd);
 
-                updateLastModified(exclude);
-                updateLastModified(composite);
+                        updateInclude = true;
+                    }
+                    //add to exclude
+                    WsAddMemberResults addMemberResults = gf.makeWsAddMemberResults(exclude, user, userToAdd);
 
-                return makeGroupingsServiceResult(addMemberResults, action);
+                    updateExclude = true;
+
+                    gsrs.add(makeGroupingsServiceResult(addMemberResults, action));
+                }
+                //They are already in the group, so just return SUCCESS
+                gsrs.add(makeGroupingsServiceResult(SUCCESS + ": " + userToAdd + " was already in " + group, action));
+
             }
+            //if owners check to see if the user is already in owners
+            else if (group.endsWith(OWNERS)) {
+                //check to see if userToAdd is already in owners
+                if (!inGroup(owners, userToAdd)) {
+                    //add userToAdd to owners
+                    WsAddMemberResults addMemberResults = gf.makeWsAddMemberResults(owners, user, userToAdd);
+
+                    updateOwners = true;
+
+                    gsrs.add(makeGroupingsServiceResult(addMemberResults, action));
+                }
+                //They are already in the group, so just return SUCCESS
+                gsrs.add(makeGroupingsServiceResult(SUCCESS + ": " + userToAdd + " was already in " + group, action));
+            }
+            //Owners can only change include, exclude and owners groups
             else {
-                return makeGroupingsServiceResult(SUCCESS + ": " + userToAdd + " was not in basis", action);
+                gsrs.add(makeGroupingsServiceResult(FAILURE + ": " + username + " may only add to exclude, include or owner group", action));
             }
-        } else if (group.endsWith(OWNERS)) {
-            WsAddMemberResults addMemberResults = gf.makeWsAddMemberResults(group, user, userToAdd);
 
-            updateLastModified(group);
-
-            return makeGroupingsServiceResult(addMemberResults, action);
+            //update groups that were changed
+            if (updateExclude) {
+                updateLastModified(exclude);
+                updateComposite = true;
+            }
+            if (updateInclude) {
+                updateLastModified(include);
+                updateComposite = true;
+            }
+            if (updateComposite) {
+                updateLastModified(composite);
+            }
+            if (updateOwners) {
+                updateLastModified(owners);
+            }
+        } else {
+            gsrs.add(makeGroupingsServiceResult(FAILURE + ": " + username + "does not have permission to edit " + group, action));
         }
-        return makeGroupingsServiceResult(FAILURE + ": " + username + " may only add to exclude, include or owner group", action);
+
+        return gsrs;
     }
 
     /**
@@ -1225,69 +1267,70 @@ public class GroupingsServiceImpl implements GroupingsService {
      * @return information about success of action
      */
     @Override
-    public GroupingsServiceResult addMemberAs(String username, String group, List<String> usersToAdd) {
-        logger.info("addMemberAs; user: " + username + "; group: " + group + "; usersToAdd: " + usersToAdd + ";");
-        //todo make this more efficient
-        for(String userToAdd : usersToAdd) {
-            GroupingsServiceResult gsr = addMemberAs(username, group, userToAdd);
+    public List<GroupingsServiceResult> addMembersAs(String username, String group, List<String> usersToAdd) {
+        logger.info("addMembersAs; user: " + username + "; group: " + group + "; usersToAdd: " + usersToAdd + ";");
+        List<GroupingsServiceResult> gsrList = new ArrayList<>();
+        for (String userToAdd : usersToAdd) {
+            gsrList.addAll(addMemberAs(username, group, userToAdd));
         }
-
-        return new GroupingsServiceResult(SUCCESS, "All additions were handled successfully.");
+        return gsrList;
     }
 
     @Override
-    public GroupingsServiceResult addMemberToGrouping(String username, String groupingPath, String userToAdd) {
-        GroupingsServiceResult groupingsServiceResults;
-        String action = username + " adds " + userToAdd + " to " + groupingPath;
+    public List<GroupingsServiceResult> addMemberToGrouping(String username, String groupingPath, String userToAdd) {
+        logger.info("addMemberToGrouping; user: " + username + "; group: " + groupingPath + "; usersToAdd: " + userToAdd + ";");
 
-        //check to see that they are not in the Grouping
-        if(!inGroup(groupingPath, userToAdd)){
-            //if they are in the exclude, get them out
-            if(inGroup(groupingPath + EXCLUDE, userToAdd)) {
-               groupingsServiceResults = deleteMemberAs(username, groupingPath + EXCLUDE, userToAdd);
-            }
-            //if they are not in the exclude then they are also not in the basis, so add them to the include
-            else{
-                groupingsServiceResults = addMemberAs(username, groupingPath + INCLUDE, userToAdd);
+        List<GroupingsServiceResult> gsrs = new ArrayList<>();
+
+        String action = "add user to " + groupingPath;
+        String basis = groupingPath + BASIS;
+        String exclude = groupingPath + EXCLUDE;
+        String include = groupingPath + INCLUDE;
+
+        //check to see if they are already in the grouping
+        if (!inGroup(groupingPath, userToAdd)) {
+            //get them out of the exclude
+            gsrs.add(deleteMemberAs(username, exclude, userToAdd));
+            //only add them to the include if they are not in the basis
+            if (!inGroup(basis, userToAdd)) {
+                gsrs.addAll(addMemberAs(username, include, userToAdd));
+            } else {
+                gsrs.add(makeGroupingsServiceResult(SUCCESS + ": " + userToAdd + " was in " + basis, action));
             }
         }
-        //if they are in the Grouping, do nothing but return SUCCESS
-        else {
-            groupingsServiceResults = makeGroupingsServiceResult(SUCCESS + ": " + userToAdd + " is already in " + groupingPath, action);
-        }
+        gsrs.add(makeGroupingsServiceResult(SUCCESS + ": " + userToAdd + " was already in " + groupingPath, action));
 
-        return groupingsServiceResults;
+        return gsrs;
     }
 
     @Override
-    public GroupingsServiceResult deleteMemberFromGrouping(String username, String groupingPath, String userToDelete) {
+    public List<GroupingsServiceResult> deleteMemberFromGrouping(String username, String groupingPath, String userToDelete) {
         logger.info("deleteMemberFromGrouping; username: "
-        + username
-        + "; groupingPath: "
-        + groupingPath + "; userToDelete: "
-        + userToDelete
-        + ";");
+                + username
+                + "; groupingPath: "
+                + groupingPath + "; userToDelete: "
+                + userToDelete
+                + ";");
 
-        GroupingsServiceResult groupingsServiceResult;
+        List<GroupingsServiceResult> gsrList = new ArrayList<>();
+
         String action = username + " deletes " + userToDelete + " from " + groupingPath;
 
         //make sure userToDelete is actually in the Grouping
-        if(inGroup(groupingPath, userToDelete)) {
+        if (inGroup(groupingPath, userToDelete)) {
             //if they are in the include group, get them out
-            if (inGroup(groupingPath + INCLUDE, userToDelete)) {
-                groupingsServiceResult = deleteMemberAs(username, groupingPath + INCLUDE, userToDelete);
-            }
+            gsrList.add(deleteMemberAs(username, groupingPath + INCLUDE, userToDelete));
             //if they are not in the include group, then they are in the basis, so add them to the exclude group
-            else {
-                groupingsServiceResult = addMemberAs(username, groupingPath + EXCLUDE, userToDelete);
+            if (inGroup(groupingPath + BASIS, userToDelete)) {
+                gsrList.addAll(addMemberAs(username, groupingPath + EXCLUDE, userToDelete));
             }
         }
         //since they are not in the Grouping, do nothing, but return SUCCESS
         else {
-            groupingsServiceResult = makeGroupingsServiceResult(SUCCESS + userToDelete + " was not in " + groupingPath, action);
+            gsrList.add(makeGroupingsServiceResult(SUCCESS + userToDelete + " was not in " + groupingPath, action));
         }
 
-        return groupingsServiceResult;
+        return gsrList;
     }
 
     /**
@@ -1306,19 +1349,23 @@ public class GroupingsServiceImpl implements GroupingsService {
                 + ";");
 
         String action = "delete " + userToDelete + " from " + group;
-        WsSubjectLookup user = gf.makeWsSubjectLookup(username);
-        if (group.endsWith(EXCLUDE) || group.endsWith(INCLUDE) || group.endsWith(OWNERS)) {
-            if (inGroup(group, userToDelete)) {
-                WsDeleteMemberResults deleteMemberResults = gf.makeWsDeleteMemberResults(group, user, userToDelete);
 
-                updateLastModified(parentGroupingPath(group));
-                updateLastModified(group);
-                return makeGroupingsServiceResult(deleteMemberResults, action);
+        if (isOwner(parentGroupingPath(group), username) || isSuperuser(username)) {
+            WsSubjectLookup user = gf.makeWsSubjectLookup(username);
+            if (group.endsWith(EXCLUDE) || group.endsWith(INCLUDE) || group.endsWith(OWNERS)) {
+                if (inGroup(group, userToDelete)) {
+                    WsDeleteMemberResults deleteMemberResults = gf.makeWsDeleteMemberResults(group, user, userToDelete);
+
+                    updateLastModified(parentGroupingPath(group));
+                    updateLastModified(group);
+                    return makeGroupingsServiceResult(deleteMemberResults, action);
+                }
+
+                return makeGroupingsServiceResult(SUCCESS + ": " + username + " was not in " + group, action);
             }
-
-            return makeGroupingsServiceResult(SUCCESS + ": " + username + " was not in " + group, action);
+            return makeGroupingsServiceResult(FAILURE + ": " + username + " may only delete from exclude, include or owner group", action);
         }
-        return makeGroupingsServiceResult(FAILURE + ": " + username + " may only delete from exclude, include or owner group", action);
+        return makeGroupingsServiceResult(FAILURE + ": " + username + " does not have permission to edit " + group, action);
     }
 
     /**
